+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2021-04-27 [0.2]
+
+- Added suport for individual control of RGB LED's through firmware version 2.00
+
 ## 2020-06-22 [0.1.4]
 
 - Added PyPi support